requirements.txt
anyio==3.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
coverage==7.2.7
exceptiongroup==1.1.1
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
ipaddress==1.0.23
packaging==23.1
pluggy==1.0.0
pydantic==1.10.8
pytest==7.3.2
pytest-cov==4.1.0
pytest-mock==3.11.1
PyYAML==6.0
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.15.0
urllib3==2.0.3
uvicorn==0.22.0
python-statemachine==2.5.0
numpy==2.2.4
locust==2.32.3
watchdog==3.0.0
transformers==4.56.2
torch==2.8.0

